Refactoring: push business logic and custom UI into their own containers
