+++ bigbluebutton-html5/imports/ui/components/notes/component.jsx
@@ -4,6 +4,7 @@ import { defineMessages, injectIntl } from 'react-intl';
 import injectWbResizeEvent from '/imports/ui/components/presentation/resize-wrapper/component';
 import Service from '/imports/ui/components/notes/service';
 import PadContainer from '/imports/ui/components/pads/container';
+import ConverterButtonContainer from './converter-button/container';
 import Styled from './styles';
 import { PANELS, ACTIONS } from '../layout/enums';
 import browserInfo from '/imports/utils/browserInfo';
@@ -17,10 +18,6 @@ const intlMessages = defineMessages({
     id: 'app.notes.title',
     description: 'Title for the shared notes',
   },
-  convertAndUploadLabel: {
-    id: 'app.note.convertAndUpload',
-    description: 'Export shared notes as a PDF and upload to the main room',
-  },
 });
 
 const propTypes = {
@@ -31,7 +28,6 @@ const propTypes = {
   hasPermission: PropTypes.bool.isRequired,
   isResizing: PropTypes.bool.isRequired,
   layoutContextDispatch: PropTypes.func.isRequired,
-  amIPresenter: PropTypes.bool.isRequired,
 };
 
 const Notes = ({
@@ -39,49 +35,10 @@ const Notes = ({
   intl,
   isRTL,
   layoutContextDispatch,
-  amIPresenter,
   isResizing,
 }) => {
   useEffect(() => () => Service.setLastRev(), []);
   const { isChrome } = browserInfo;
-
-  if (hasPermission && amIPresenter) {
-    return (
-      <Styled.Notes data-test="notes" isChrome={isChrome}>
-        <Styled.Header>
-          <Styled.Title data-test="notesTitle">
-            <Styled.HideButton
-              onClick={() => {
-                layoutContextDispatch({
-                  type: ACTIONS.SET_SIDEBAR_CONTENT_IS_OPEN,
-                  value: false,
-                });
-                layoutContextDispatch({
-                  type: ACTIONS.SET_SIDEBAR_CONTENT_PANEL,
-                  value: PANELS.NONE,
-                });
-              }}
-              data-test="hideNotesLabel"
-              aria-label={intl.formatMessage(intlMessages.hide)}
-              label={intl.formatMessage(intlMessages.title)}
-              icon={isRTL ? 'right_arrow' : 'left_arrow'}
-            />
-          </Styled.Title>
-          <Styled.ConvertAndUpload
-            onClick={() => { Service.convertAndUpload(Date.now()); }}
-            label={intl.formatMessage(intlMessages.convertAndUploadLabel)}
-            icon="upload"
-          />
-        </Styled.Header>
-        <PadContainer
-          externalId={Service.ID}
-          hasPermission={hasPermission}
-          isResizing={isResizing}
-          isRTL={isRTL}
-        />
-      </Styled.Notes>
-    );
-  }
 
   return (
     <Styled.Notes data-test="notes" isChrome={isChrome}>
@@ -104,6 +61,7 @@ const Notes = ({
             icon={isRTL ? 'right_arrow' : 'left_arrow'}
           />
         </Styled.Title>
+        <ConverterButtonContainer />
       </Styled.Header>
       <PadContainer
         externalId={Service.ID}
