+++ bigbluebutton-html5/imports/ui/components/pads/service.js
@@ -2,10 +2,8 @@ import Pads, { PadsUpdates } from '/imports/api/pads';
 import { makeCall } from '/imports/ui/services/api';
 import Auth from '/imports/ui/services/auth';
 import Settings from '/imports/ui/services/settings';
-import PresentationUploaderService from '/imports/ui/components/presentation/presentation-uploader/service';
 
 const PADS_CONFIG = Meteor.settings.public.pads;
-const PRESENTATION_CONFIG = Meteor.settings.public.presentation;
 
 const getLang = () => {
   const { locale } = Settings.application;
@@ -87,40 +85,6 @@ const getPadContent = (externalId) => {
   return '';
 };
 
-async function convertAndUpload(externalId) {
-  const params = getParams();
-  const padId = await makeCall('getPadId', externalId);
-
-  const exportUrl = Auth.authenticateURL(`${PADS_CONFIG.url}/p/${padId}/export/pdf?${params}`);
-  const sharedNotesAsPdf = await fetch(exportUrl, { credentials: 'include' });
-  const data = await sharedNotesAsPdf.blob();
-
-  const sharedNotesData = new File([data], 'SharedNotes.pdf', {
-    type: data.type,
-  });
-
-  const formData = new FormData();
-
-  formData.append('presentation_name', 'SharedNotes.pdf');
-  formData.append('Filename', 'SharedNotes.pdf');
-  formData.append('conference', Auth.meetingID);
-  formData.append('room', Auth.meetingID);
-  formData.append('pod_id', 'DEFAULT_PRESENTATION_POD');
-  formData.append('is_downloadable', false);
-  formData.append('current', true);
-  formData.append('fileUpload', sharedNotesData);
-
-  const presentationUploadToken = await PresentationUploaderService.requestPresentationUploadToken('DEFAULT_PRESENTATION_POD', Auth.meetingID, 'SharedNotes.pdf');
-
-  fetch(PRESENTATION_CONFIG.uploadEndpoint.replace('upload', `${presentationUploadToken}/upload`), {
-    body: formData,
-    method: 'post',
-  });
-
-  makeCall('setUsedToken', presentationUploadToken);
-  return null;
-}
-
 export default {
   getPadId,
   createGroup,
@@ -130,5 +94,5 @@ export default {
   getRev,
   getPadTail,
   getPadContent,
-  convertAndUpload,
+  getParams,
 };
