+++ bigbluebutton-html5/public/locales/en.json
@@ -58,7 +58,7 @@
     "app.notes.title": "Shared Notes",
     "app.notes.label": "Notes",
     "app.notes.hide": "Hide notes",
-    "app.note.convertAndUpload": "Move notes to whiteboard",
+    "app.note.converter-button.convertAndUpload": "Move notes to whiteboard",
     "app.notes.locked": "Locked",
     "app.pads.hint": "Press Esc to focus the pad's toolbar",
     "app.user.activityCheck": "User activity check",
